requirements.txt
aiohttp==3.9.3
aiokafka==0.10.0
aiosignal==1.3.1
async-timeout==4.0.3
asyncio-dgram==2.1.2
attrs==23.2.0
discord.py==2.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
mcstatus==11.1.1
multidict==6.0.5
packaging==23.2
python-dotenv==1.0.1
setuptools==68.2.2
wheel==0.41.2
yarl==1.9.4
